Groups and tasklists are now only accessible for authorized users. Only group admins are able to edit group info and group member roles

+++ src/main/java/nl/miwgroningen/se6/heartcoded/CaTo/controller/GroupController.java
@@ -52,6 +52,11 @@ public class GroupController {
 
     @GetMapping("/groups/delete/{groupId}")
     protected String deleteGroupBySiteAdmin(@PathVariable("groupId") Integer groupId) {
+
+        if (!memberService.userIsMemberOfGroup(groupId) || !memberService.userIsGroupAdmin(groupId)) {
+            return "redirect:/403";
+        }
+
         taskListService.deleteByGroupId(groupId);
         groupService.deleteGroupById(groupId);
         return "redirect:/groups";
